pyproject.toml
[project]
name = "overhearops"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "langgraph>=1.0.0a0",
  "langgraph-checkpoint-sqlite>=2.0.11",
  "opentelemetry-sdk>=1.26",
  "opentelemetry-exporter-otlp>=1.26",
  "httpx>=0.27",
  "rich>=13.7",
  "pytest>=8.3",
  "ruff>=0.6.4",
  "mypy>=1.11",
  "python-dotenv>=1.0",
  "orjson>=3.10",
  "honcho>=1.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["apps", "packages"]
include = ["service*", "agentkit*", "obs*"]

[tool.uv]
package = true
prerelease = "allow"

[tool.pytest.ini_options]
addopts = "-ra"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "ASYNC", "S"]
ignore = ["S101"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
strict_optional = true
mypy_path = ["packages", "apps"]

[tool.coverage.run]
branch = true
source = ["apps", "packages"]
